Be more careful about casting and integral types
 - Introduce safe casting functions which assert if precision is lost

 - Use safe casting functions to do conversions

 - Use `Eigen::Index` for indexing Eigen types and `std::size_t` for
   indexing STL types

 - A few other small changes to e.g. function argument types where
    things didn't line up

+++ include/albatross/src/core/parameters.hpp
@@ -103,7 +103,8 @@ inline std::string pretty_param_details(const ParameterStore &params) {
   for (const auto &pair : params) {
     std::string prior_name;
     prior_name = pair.second.prior.get_name();
-    ss << "    " << std::left << std::setw(max_name_length + 1) << pair.first
+    ss << "    " << std::left
+       << std::setw(static_cast<int>(max_name_length + 1)) << pair.first
        << " value: " << std::left << std::setw(12) << pair.second.value
        << " valid: " << std::left << std::setw(3) << pair.second.is_valid()
        << " prior: " << std::setw(15) << prior_name << " bounds: ["
